# expression: implement vectorized evaluation for `builtinIn*Sig`