--- a/expression/generator/other_vec.go
+++ b/expression/generator/other_vec.go
@@ -124,6 +124,9 @@ func (b *{{.SigName}}) vecEvalInt(input *chunk.Chunk, result *chunk.Column) erro
 	{{- end }}
 	result.ResizeInt64(n, true)
 	r64s := result.Int64s()
+	for i:=0; i<n; i++ {
+		r64s[i] = 0
+	}
 	hasNull := make([]bool, n)
 	{{- if $InputInt }}
 		isUnsigned0 := mysql.HasUnsignedFlag(b.args[0].GetType().Flag)